New Development: No             JIRA Issue: CAS-4106         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: test_partition, manual tests   Put in Release Notes: No              Module(s): gcwrap            Description: Restored a feature which got lost in 			my previous change: master-subMS should 			have a full copy of the POINTING and SYSCAL 			table under all circumstances.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@21098 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ gcwrap/python/scripts/task_partition.py
@@ -63,16 +63,18 @@ class PartitionHelper(ParallelTaskHelper):
             mytb = tbtool()
             mytb.open(self.ptab)
             self.pointingisempty = (mytb.nrows()==0)
-            self.makepointinglinks = not self.pointingisempty
             mytb.close()
+            self.makepointinglinks = not self.pointingisempty
+            self.pwriteaccess = True
             
             self.syscalisempty = True
             self.makesyscallinks = False
+            self.swriteaccess = True
             if(os.path.exists(self.stab)): # syscal is optional
                 mytb.open(self.stab)
                 self.syscalisempty = (mytb.nrows()==0)
-                self.makesyscallinks = not self.syscalisempty
                 mytb.close()
+                self.makesyscallinks = not self.syscalisempty
 
             if not self.pointingisempty:
                 if os.access(os.path.dirname(self.ptab), os.W_OK):
@@ -84,7 +86,7 @@ class PartitionHelper(ParallelTaskHelper):
                     tmpp.close()
                     mytb.close()
                 else:
-                    self.pointingisempty = True # effectively
+                    self.pwriteaccess = False
                     
 
             if not self.syscalisempty:
@@ -97,7 +99,7 @@ class PartitionHelper(ParallelTaskHelper):
                     tmpp.close()
                     mytb.close()
                 else:
-                    self.syscalisempty = True # effectively
+                    self.swriteaccess = False
 
 
     def generateJobs(self):
@@ -342,11 +344,11 @@ class PartitionHelper(ParallelTaskHelper):
             casalog.post("Finalizing MMS structure")
 
             # restore POINTING and SYSCAL
-            if not self.pointingisempty:
+            if self.pwriteaccess and not self.pointingisempty:
                 print "restoring POINTING"
                 os.system('rm -rf '+self.ptab) # remove empty copy
                 os.system('mv '+self.dataDir+'/POINTING '+self.ptab)
-            if not self.syscalisempty:
+            if self.swriteaccess and not self.syscalisempty:
                 print "restoring SYSCAL"
                 os.system('rm -rf '+self.stab) # remove empty copy
                 os.system('mv '+self.dataDir+'/SYSCAL '+self.stab)
